bsdsocket: declare gethostname with size_t length, per POSIX
The historical int declaration conflicted with POSIX declarations
(e.g. gnulib's unistd.h) in code seeing both. Register-compatible
with the previous signature on all targets.

+++ workbench/network/common/include/clib/bsdsocket_protos.h
@@ -254,7 +254,7 @@ AROS_LP3(int, recvmsg,
 );
 AROS_LP2(int, gethostname,
          AROS_LPA(char *, name, A0),
-         AROS_LPA(int, namelen, D0),
+         AROS_LPA(size_t, namelen, D0),
          LIBBASETYPEPTR, SocketBase, 47, BSDSocket
 );
 AROS_LP0(long, gethostid,
